+++ examples/worklet.rs
@@ -5,7 +5,7 @@ use web_audio_api::context::{
 };
 use web_audio_api::node::{AudioNode, ChannelConfig, ChannelConfigOptions};
 use web_audio_api::param::{AudioParam, AudioParamOptions, AutomationRate};
-use web_audio_api::renderer::{AudioParamValues, AudioProcessor, AudioRenderQuantum};
+use web_audio_api::render::{AudioParamValues, AudioProcessor, AudioRenderQuantum};
 use web_audio_api::SampleRate;
 
 /// Audio source node emitting white noise (random samples)
