+++ src/lib.rs
@@ -55,7 +55,7 @@ pub mod control;
 pub mod media;
 pub mod node;
 pub mod param;
-pub mod renderer;
+pub mod render;
 pub mod spatial;
 
 #[cfg(test)]
