+++ src/node/destination.rs
@@ -1,5 +1,5 @@
 use crate::context::{AsBaseAudioContext, AudioContextRegistration};
-use crate::renderer::{AudioParamValues, AudioProcessor, AudioRenderQuantum};
+use crate::render::{AudioParamValues, AudioProcessor, AudioRenderQuantum};
 use crate::SampleRate;
 
 use super::{
